requirements.txt
numpy
mlxtend
matplotlib
pandas
scikit-learn
Tensorflow-gpu
keras
